fix terminal name retrieval and update packages
despite convention, TERM_PROGRAM is only set on some terminals and the unwrap() leads to crashes when ran terminals that don't set it, such as Alacritty and kitty (#2). This issue is fixed by checking for terminal-specific vars for these two terminals and adds a graceful "unknown" fallback option in case anyone dare use a different terminal emulator.

Also had to run cargo update to make the program compile on the newest rust version.